5.0.13 version commit
git-svn-id: https://svn.ufz.de/svn/ogs/trunk/sources@7608 3895a583-e809-0410-9bb6-aa9463026377

+++ FEM/par_ddc.cpp
@@ -823,7 +823,7 @@ void CPARDomain::InitialEQS(CRFProcess *m_pcs)
    }
    else
       this_eqs = eqs;
-   for(int i=0; i<m_pcs->GetPrimaryVNumber(); i++)
+   for(size_t i=0; i<m_pcs->GetPrimaryVNumber(); i++)
       shift[i] = i*size;
    //
    this_eqs->Initialize();
